Buy page
why the buy page shows no results

Because there are none, and we have nothing special that prints when no books are entered into the system. So we need to insert a link or something if there's no options which sends you to Amazon or ebay or vezzi or something.