Books now have owner classes
You can reach the email to contact the book's owner now. It's in the open - I'm not bothering to cover it right now.

Login/logout looks better, although I still haven't tested it enough to know. I'm still working on that.

I haven't got the images yet, and I don't know if I'll get to it. However, I got the user-email & book to use the keyproperty, so i have a chance.